feat: Separate raw conversions
