+++ DUI3-DX/Converters/Rhino/Speckle.Converters.Rhino7/ToSpeckle/Raw/IntervalToSpeckleConverter.cs
@@ -4,10 +4,7 @@ using Speckle.Core.Models;
 
 namespace Speckle.Converters.Rhino7.Geometry;
 
-[NameAndRankValue(nameof(RG.Interval), NameAndRankValueAttribute.SPECKLE_DEFAULT_RANK)]
-public class IntervalToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConversion<RG.Interval, SOP.Interval>
+public class IntervalToSpeckleConverter : IRawConversion<RG.Interval, SOP.Interval>
 {
-  public Base Convert(object target) => RawConvert((RG.Interval)target);
-
   public SOP.Interval RawConvert(RG.Interval target) => new(target.T0, target.T1);
 }
