+++ DUI3-DX/Converters/Rhino/Speckle.Converters.Rhino7/ToSpeckle/Raw/PointToSpeckleConverter.cs
@@ -1,13 +1,11 @@
 using Rhino;
-using Speckle.Converters.Common.Objects;
 using Speckle.Converters.Common;
-using Speckle.Core.Models;
+using Speckle.Converters.Common.Objects;
 
-namespace Speckle.Converters.Rhino7.Geometry;
+namespace Speckle.Converters.Rhino7.ToSpeckle.Raw;
 
 // POC: not sure I like the place of the default rank
-[NameAndRankValue(nameof(RG.Point3d), NameAndRankValueAttribute.SPECKLE_DEFAULT_RANK)]
-public class PointToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConversion<RG.Point3d, SOG.Point>
+public class PointToSpeckleConverter : IRawConversion<RG.Point3d, SOG.Point>
 {
   private readonly IConversionContextStack<RhinoDoc, UnitSystem> _contextStack;
 
@@ -15,8 +13,6 @@ public class PointToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConve
   {
     _contextStack = contextStack;
   }
-
-  public Base Convert(object target) => RawConvert((RG.Point3d)target);
 
   public SOG.Point RawConvert(RG.Point3d target) =>
     new(target.X, target.Y, target.Z, _contextStack.Current.SpeckleUnits);
