+++ DUI3-DX/Converters/Rhino/Speckle.Converters.Rhino7/ToSpeckle/Raw/VectorToSpeckleConverter.cs
@@ -1,12 +1,10 @@
 ﻿using Rhino;
 using Speckle.Converters.Common;
 using Speckle.Converters.Common.Objects;
-using Speckle.Core.Models;
 
-namespace Speckle.Converters.Rhino7.Geometry;
+namespace Speckle.Converters.Rhino7.ToSpeckle.Raw;
 
-[Common.NameAndRankValue(nameof(RG.Vector3d), NameAndRankValueAttribute.SPECKLE_DEFAULT_RANK)]
-public class VectorToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConversion<RG.Vector3d, SOG.Vector>
+public class VectorToSpeckleConverter : IRawConversion<RG.Vector3d, SOG.Vector>
 {
   private readonly IConversionContextStack<RhinoDoc, UnitSystem> _contextStack;
 
@@ -14,8 +12,6 @@ public class VectorToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConv
   {
     _contextStack = contextStack;
   }
-
-  public Base Convert(object target) => RawConvert((RG.Vector3d)target);
 
   public SOG.Vector RawConvert(RG.Vector3d target) =>
     new(target.X, target.Y, target.Z, _contextStack.Current.SpeckleUnits);
